# Conflicts:
#	.laya/web.json
#	release/web/index-4362419498.js
#	release/web/index-6f0f4d6803.js
#	release/web/index-7b8b9dc1c5.js
#	release/web/index.html
#	release/web/version-4d9a9511c2.json
#	release/web/version-7668d4e729.json
#	release/web/version-899641e600.json
#	src/view/FriendsView.ts

+++ libs/hd/createRes.js
@@ -1,8 +1,7 @@
 const chokidar = require("chokidar");
 const fs = require("fs");
 const path = require("path");
-
-const buildType = process.argv[2];
+const { buildType } = require("./config");
 
 /////======================================
 
@@ -95,7 +94,10 @@ export default Res;
             return;
         }
         console.info("\x1b[33m", "更新生成 Res.ts");
-        if (!process.argv.includes("debug")) {
+        if (
+            !process.argv.includes(buildType.debug) &&
+            !process.argv.includes(buildType.debug_test_start)
+        ) {
             console.log("结束构建");
             process.kill(process.pid);
         }
